asks for confirmation when deleting a non-empty list (with javascript confirm, boostrap modal would look better)

+++ jsx/components/memolist.jsx
@@ -41,12 +41,20 @@ class Memolist extends React.Component{
     handleDeleteClick = () => {
       const list_index = this.props.list_index;
       const uri = this.props.lists[list_index].uri;
-      this.props.delete_list_remote(list_index, uri);
+      if (this.props.lists[list_index].items.length == 0 || confirm("This list contains items.\nDelete List?")) {
+        this.props.delete_list_remote(list_index, uri);
+      }
     }
 
     handlePDFClick = (event) => {
       event.preventDefault();
       const list_index = this.props.list_index;
+
+      /*
+        Check if XMLHttpRequest is available and build PDF.
+        Mocha tests are run outside the browser, required to exclude this
+        from testing (which would fail).
+      */
       if (window.XMLHttpRequest !== undefined)
       {
           var jsPDF = require('../external/jspdf.debug');
@@ -105,8 +113,10 @@ class Memolist extends React.Component{
                  <div className="pull-right">
                         <ButtonToolbar>
                         <DropdownButton bsStyle="link" title={ dropdown_button } id="dropdown-link" noCaret pullRight style={ dropdown_button_style }>
-                        <MenuItem eventKey="1"><span onClick={ this.handlePDFClick } className="text-muted"><span className="glyphicon glyphicon-download-alt" aria-hidden="true"></span> Get PDF</span></MenuItem>
-                        <MenuItem eventKey="2"><span onClick={ this.handleDeleteClick } className="text-muted"><span className="glyphicon glyphicon-remove" aria-hidden="true"></span> Delete list</span></MenuItem>
+                        <MenuItem header>Actions</MenuItem>
+                        <MenuItem eventKey="1" onClick={ this.handlePDFClick }><span className="text-muted"><span className="glyphicon glyphicon-download-alt" aria-hidden="true"></span> Get PDF</span></MenuItem>
+                        <MenuItem divider/>
+                        <MenuItem eventKey="2" onClick={ this.handleDeleteClick }><span className="text-muted"><span className="glyphicon glyphicon-remove" aria-hidden="true"></span> Delete list</span></MenuItem>
                         </DropdownButton>
                     </ButtonToolbar>
                  </div>
